refactor(core): split internals by responsibility (#26)
* refactor(core): split internals by responsibility

* refactor(core): introduce analyzer service classes

* refactor(core): add shared analyzer base class

* refactor(app): add mode runner template method

* refactor(app): move mode runners into dedicated modules

* refactor(app): rename runners by domain

* refactor(app): rename runners to commands

* refactor(output): group ascii and json outputs

* refactor(api): remove re-export facades

* refactor(types): split types into dedicated modules

* refactor(types): remove type barrel exports
