+++ src/index.ts
@@ -1,27 +1,28 @@
-export { analyzeDependencies, graphToSerializableTree } from './analyzer.js'
+export { analyzeDependencies } from './analyzers/import/index.js'
+export { analyzeReactUsage } from './analyzers/react/index.js'
 export {
-  analyzeReactUsage,
+  getFilteredUsages,
+  getReactUsageEntries,
   getReactUsageRoots,
-  graphToSerializableReactTree,
-} from './react-analyzer.js'
-export { printReactUsageTree } from './react-tree.js'
-export { printDependencyTree } from './tree.js'
-export type {
-  AnalyzeOptions,
-  ColorMode,
-  DependencyEdge,
-  DependencyGraph,
-  DependencyKind,
-  PrintReactTreeOptions,
-  PrintTreeOptions,
-  ReactSymbolKind,
-  ReactUsageEdge,
-  ReactUsageEdgeKind,
-  ReactUsageEntry,
-  ReactUsageFilter,
-  ReactUsageGraph,
-  ReactUsageLocation,
-  ReactUsageNode,
-  ReferenceKind,
-  SourceModuleNode,
-} from './types.js'
+} from './analyzers/react/queries.js'
+export { printDependencyTree } from './output/ascii/import.js'
+export { printReactUsageTree } from './output/ascii/react.js'
+export { graphToSerializableTree } from './output/json/import.js'
+export { graphToSerializableReactTree } from './output/json/react.js'
+export type { AnalyzeOptions } from './types/analyze-options.js'
+export type { ColorMode } from './types/color-mode.js'
+export type { DependencyEdge } from './types/dependency-edge.js'
+export type { DependencyGraph } from './types/dependency-graph.js'
+export type { DependencyKind } from './types/dependency-kind.js'
+export type { PrintReactTreeOptions } from './types/print-react-tree-options.js'
+export type { PrintTreeOptions } from './types/print-tree-options.js'
+export type { ReactSymbolKind } from './types/react-symbol-kind.js'
+export type { ReactUsageEdge } from './types/react-usage-edge.js'
+export type { ReactUsageEdgeKind } from './types/react-usage-edge-kind.js'
+export type { ReactUsageEntry } from './types/react-usage-entry.js'
+export type { ReactUsageFilter } from './types/react-usage-filter.js'
+export type { ReactUsageGraph } from './types/react-usage-graph.js'
+export type { ReactUsageLocation } from './types/react-usage-location.js'
+export type { ReactUsageNode } from './types/react-usage-node.js'
+export type { ReferenceKind } from './types/reference-kind.js'
+export type { SourceModuleNode } from './types/source-module-node.js'
