+++ src/types/analyze-options.ts
@@ -1,0 +1,4 @@
+export interface AnalyzeOptions {
+  readonly cwd?: string
+  readonly configPath?: string
+}
