+++ src/types/print-react-tree-options.ts
@@ -1,0 +1,8 @@
+import type { ColorMode } from './color-mode.js'
+import type { ReactUsageFilter } from './react-usage-filter.js'
+
+export interface PrintReactTreeOptions {
+  readonly cwd?: string
+  readonly filter?: ReactUsageFilter
+  readonly color?: ColorMode
+}
