+++ src/types/react-usage-edge-kind.ts
@@ -1,0 +1,1 @@
+export type ReactUsageEdgeKind = 'render' | 'hook-call'
