+++ src/types/react-usage-entry.ts
@@ -1,0 +1,6 @@
+import type { ReactUsageLocation } from './react-usage-location.js'
+
+export interface ReactUsageEntry {
+  readonly target: string
+  readonly location: ReactUsageLocation
+}
