+++ src/types/react-usage-graph.ts
@@ -1,0 +1,9 @@
+import type { ReactUsageEntry } from './react-usage-entry.js'
+import type { ReactUsageNode } from './react-usage-node.js'
+
+export interface ReactUsageGraph {
+  readonly cwd: string
+  readonly entryId: string
+  readonly nodes: ReadonlyMap<string, ReactUsageNode>
+  readonly entries: readonly ReactUsageEntry[]
+}
