+++ src/types/reference-kind.ts
@@ -1,0 +1,6 @@
+export type ReferenceKind =
+  | 'import'
+  | 'export'
+  | 'require'
+  | 'dynamic-import'
+  | 'import-equals'
