+++ src/types/source-module-node.ts
@@ -1,0 +1,6 @@
+import type { DependencyEdge } from './dependency-edge.js'
+
+export interface SourceModuleNode {
+  readonly id: string
+  readonly dependencies: readonly DependencyEdge[]
+}
