+++ src/utils/is-source-code-file.ts
@@ -1,0 +1,16 @@
+import path from 'node:path'
+
+export const SOURCE_EXTENSIONS = new Set([
+  '.js',
+  '.jsx',
+  '.ts',
+  '.tsx',
+  '.mjs',
+  '.cjs',
+  '.mts',
+  '.cts',
+])
+
+export function isSourceCodeFile(filePath: string): boolean {
+  return SOURCE_EXTENSIONS.has(path.extname(filePath).toLowerCase())
+}
