+++ src/utils/normalize-file-path.ts
@@ -1,0 +1,5 @@
+import path from 'node:path'
+
+export function normalizeFilePath(filePath: string): string {
+  return path.normalize(filePath)
+}
